% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_lasso_reg.R
\name{pool_lasso_reg}
\alias{pool_lasso_reg}
\title{Use pool lasso regression to do variable selection,
If a variable has two very stong lags, then delete one of these 2 lag to let more variables in}
\usage{
pool_lasso_reg(dep, actual, w, date.var, date.start, date.end, group, data,
  indepvar, is.plot = T, is.weights = T, is.intercept = T,
  is.multithread = F, spec = rep("localhost", 4))
}
\description{
Use pool lasso regression to do variable selection,
If a variable has two very stong lags, then delete one of these 2 lag to let more variables in
}

